Cleanup pass on TestBrowser.
