+++ osu.Framework.Testing/osu.Framework.Testing.csproj
@@ -47,7 +47,9 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestBrowser.cs" />
+    <Compile Include="TestBrowserConfig.cs" />
     <Compile Include="TestCase.cs" />
+    <Compile Include="Drawables\TestCaseButton.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\osu.Framework\osu.Framework.csproj">
